feat: moving shadowdom funcs to utilties package
